feat: enable checkpointing (#19)
* feat: simple checkpointing

Works only for feature extractor and regressor. Option to add other networks, like domain discriminator, still missing.

* feat: extended checkpointing

Works with arbitrary number of models. They need to be mentioned in the CHECKPOINT_MODEL class variable, though.

* feat: throw error if model it not checkpoint-able

* feat: make nested models checkpoint-able

Nested models take another nn.Module as an init param. This init module should not be pickled directly but saved the same way as the top-level module. An example for a nested model is the DannLoss that takes a domain disc.

* docs: add docstring

* fix: linting issues